member account
Former-commit-id: 22e3a3823f9f74e3377cff3db76386576388070a

+++ core/domain/interface/payment/payment.go
@@ -117,7 +117,7 @@ type (
 		BalanceDiscount(remark string) error
 
 		// 使用会员积分抵扣,返回抵扣的金额及错误,ignoreOut:是否忽略超出订单金额的积分
-		IntegralDiscount(integral int, ignoreOut bool) (float32, error)
+		IntegralDiscount(integral int64, ignoreOut bool) (float32, error)
 
 		// 系统支付金额
 		SystemPayment(fee float32) error
